+++ src/core/control/tools/Selection.cpp
@@ -20,7 +20,7 @@ auto Selection::finalize(PageRef page) -> size_t {
     size_t layerId = 0;
 
     if (multiLayer) {
-        for (int layerNo = page->getLayers()->size() - 1; layerNo >= 0; layerNo--) {
+        for (int layerNo = static_cast<int>(page->getLayers()->size()) - 1; layerNo >= 0; layerNo--) {
             Layer* l = page->getLayers()->at(layerNo);
             if (!l->isVisible()) {
                 continue;
